move Test_isStructured to requirements package
